@@ -1,36 +1,21 @@
 {
-    "name": "Bootwind-Template",
-    "version": "0.0.2",
-    "private": true,
-    "scripts": {
-        "serve": "vue-cli-service serve",
-        "build": "vue-cli-service build"
-    },
-    "dependencies": {
-        "apexcharts": "^3.19.2",
-        "register-service-worker": "^1.7.1",
-        "vee-validate": "^3.3.5",
-        "vue": "^2.6.11",
-        "vue-apexcharts": "^1.5.3",
-        "vue-i18n": "^8.18.2",
-        "vue-prism-editor": "^0.6.1",
-        "vue-router": "^3.2.0",
-        "vue-scrollto": "^2.18.1",
-        "vuedraggable": "^2.23.2",
-        "vuex": "^3.4.0"
-    },
-    "devDependencies": {
-        "@vue/cli-plugin-pwa": "^4.4.0",
-        "@vue/cli-plugin-router": "^4.4.0",
-        "@vue/cli-plugin-vuex": "^4.4.0",
-        "@vue/cli-service": "^4.4.0",
-        "sass-loader": "^8.0.2",
-        "vue-cli-plugin-tailwind": "^1.4.1",
-        "vue-template-compiler": "^2.6.11"
-    },
-    "browserslist": [
-        "> 1%",
-        "last 2 versions",
-        "not dead"
-    ]
+  "name": "bootwind",
+  "private": true,
+  "version": "0.0.3",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
+  },
+  "dependencies": {
+    "vue": "^3.3.4"
+  },
+  "devDependencies": {
+    "@vitejs/plugin-vue": "^4.2.3",
+    "autoprefixer": "^10.4.16",
+    "postcss": "^8.4.30",
+    "tailwindcss": "^3.3.3",
+    "vite": "^4.4.5"
+  }
 }
