@@ -1,14 +1,5 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
+import './style.css'
 import App from './App.vue'
-import './registerServiceWorker'
-import router from './router'
-import store from './store'
-import './assets/tailwind.css'
 
-
-Vue.config.productionTip = false
-new Vue({
-    router,
-    store,
-    render: function(h) { return h(App) }
-}).$mount('#app')
+createApp(App).mount('#app')
